pyproject.toml
[project]
authors = [
  {name = "Saito Tsutomu", email = "tsutomu7@hotmail.co.jp"},
]
classifiers = [
  "Development Status :: 1 - Planning",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Topic :: Software Development",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "nicegui>=2.11.1",
  "pytest>=8.3.4",
  "tortoise-orm>=0.24.0",
]
description = "Book list sample application on tortoise database"
name = "nicegui-book-list"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[project.urls]
docs = "https://saitotsutomu.github.io/nicegui-book_list/"
homepage = "https://github.com/SaitoTsutomu/nicegui-book_list"

[project.scripts]
book-list = "nicegui_book_list:main"

[tool.uv]
dev-dependencies = []

[tool.mypy]
files = "src"
ignore_missing_imports = true

[tool.ruff]
src = ["src"]

[tool.ruff.lint]
ignore = ["CPY", "D203", "D213", "D400", "D415", "DOC", "E501"]
preview = true
select = ["ALL"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]
